move classes to package level and add docs
